כל התאריכים באפליקציה בלוח העברי (גימטריה) + רמזי-תאריך עברי בשדות
- formatHebrewDate/formatHebrewDateTime עברו ללוח העברי (he-IL-u-ca-hebrew):
  שם-חודש עברי מ-Intl (כולל אדר א׳/ב׳ בשנה מעוברת), ויום+שנה מומרים
  לגימטריה עם גרש/גרשיים (כ״ז בתמוז תשפ״ו). מכיוון שכמעט כל האפליקציה
  כבר עוברת דרך העוזרים האלה — כל התצוגות התחלפו ללוח העברי בבת-אחת.
- תוויות התאריך בגרפים (useDashboardData) והדוחות (Lateness/Absence)
  יושרו ללוח העברי עם שמות-חודשים.
- שדות <input type="date"> נשארים לועזיים (נדרש ISO), אך נוסף מתחתם
  רמז המציג את התאריך העברי: תאריך הדוח בהעלאה, וטווח מתאריך/עד-תאריך
  בעמוד הדוחות.

+++ src/components/reports/AbsenceReport.tsx
@@ -31,7 +31,7 @@ const rangeSchema = z
   .refine((v) => v.from <= v.to, { message: "תאריך ההתחלה חייב להיות לפני תאריך הסיום" });
 
 function shortDate(iso: string): string {
-  return formatHebrewDate(iso, { day: "numeric", month: "numeric" });
+  return formatHebrewDate(iso, { day: "numeric", month: "short" });
 }
 
 /**
